Moved backend modules into a new hierarchy
